rss_pipeline/rag/requirements.txt
psycopg2-binary
openai